# nerdbank/openapi.json
{
    "openapi": "3.0.0",
    "info": {
    "title": "Transfers API",
    "version": "1.0.0",
    "description": "API for managing transfers"
    },
    "servers": [
    {
    "url": "https://nerdbank.onrender.com/api"
    }
    ],
    "paths": {
    "/accounts/accounts": {
    "post": {
    "summary": "Create a new account",
    "description": "Adds a new account to the system.",
    "operationId": "Createaccounts",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "properties": {
    "accountName": {
    "type": "string"
    },
    "accountType": {
    "type": "string"
    },
    "accountPin": {
    "type": "string"
    },
    "accountBankName": {
    "type": "string"
    },
    "accountBankSwiftCode": {
    "type": "string"
    },
    "accountBankRoutingTransitNumber": {
    "type": "string"
    },
    "accountNumber": {
    "type": "string"
    },
    "accountBalance": {
    "type": "string"
    },
    "accountStatus": {
    "type": "string"
    }
    }
    }
    }
    }
    },
    "responses": {
    "201": {
    "description": "accounts created successfully"
    },
    "500": {
    "description": "Server error"
    }
    }
    },
    "get": {
    "summary": "Get all accounts",
    "description": "Returns a list of accounts",
    "operationId": "Getaccounts",
    "responses": {
    "200": {
    "description": "A JSON array of accounts"
    }
    }
    }
    },
...
    "/bank/add-bank-detail": {
    "post": {
    "summary": "Add new bank detail",
    "description": "Adds a new bank detail to the user's profile.",
    "operationId": "AddBankDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "properties": {
    "typeOfBank": {
    "type": "string"
    },
    "bankCountry": {
    "type": "string"
    },
    "bankName": {
    "type": "string"
    },
    "accountName": {
    "type": "string"
    },
    "accountNumber": {
    "type": "string"
    },
    "swiftCode": {
    "type": "string"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Bank detail added successfully"
    },
    "500": {
    "description": "Server error"
    }
    }
    }
    },
...
    "/bank/delete-bank-detail": {
    "delete": {
    "summary": "Delete a specific bank detail",
    "description": "Deletes a specific bank detail for a user.",
    "operationId": "DeleteBankDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "properties": {
    "detailId": {
    "type": "string"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Bank detail deleted successfully"
    },
    "404": {
    "description": "Bank detail not found"
    }
    }
    }
    },
    "/bank/update-bank-detail": {
    "patch": {
    "summary": "Update a specific bank detail",
    "description": "Updates a specific bank detail for a user.",
    "operationId": "UpdateBankDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "required": [
    "userId",
    "updateData"
    ],
    "properties": {
    "detailId": {
    "type": "string"
    },
    "updateData": {
    "type": "object"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Bank detail updated successfully"
    },
    "404": {
    "description": "Bank detail not found"
    }
    }
    }
    },
    "/card/add-card-detail": {
    "post": {
    "summary": "Add a new card detail",
    "description": "Adds a new card detail to the user's profile.",
    "operationId": "AddCardDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "properties": {
    "card": {
    "type": "string"
    },
    "cardType": {
    "type": "string"
    },
    "cardNumber": {
    "type": "string"
    },
    "expiryDate": {
    "type": "string"
    },
    "cvv": {
    "type": "string"
    },
    "cardHolderName": {
    "type": "string"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Card detail added successfully"
    },
    "500": {
    "description": "Server error"
    }
    }
    }
    },
...
    "/card/delete-card-detail": {
    "delete": {
    "summary": "Delete a specific card detail",
    "description": "Deletes a specific card detail for a user.",
    "operationId": "DeleteCardDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "properties": {
    "detailId": {
    "type": "string"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Card detail deleted successfully"
    },
    "404": {
    "description": "Card detail not found"
    }
    }
    }
    },
    "/card/update-card-detail": {
    "patch": {
    "summary": "Update a specific card detail",
    "description": "Updates a specific card detail for a user.",
    "operationId": "UpdateCardDetail",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "object",
    "required": [
    "userId",
    "updateData"
    ],
    "properties": {
    "detailId": {
    "type": "string"
    },
    "updateData": {
    "type": "object"
    }
    }
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Card detail updated successfully"
    },
    "404": {
    "description": "Card detail not found"
    }
    }
    }
    },
...
    },
...
    }